Move functionality from timeframe to scheduler
